Add missing keyframes to countdown pill animation
