+++ ansible/inventory.ini
@@ -1,0 +1,14 @@
+[master]
+m2 ansible_host=192.168.50.6
+m1 ansible_host=192.168.50.2
+m3 ansible_host=192.168.50.3
+
+[worker]
+w2 ansible_host=192.168.50.4
+w3 ansible_host=192.168.50.5
+w5 ansible_host=192.168.50.7
+
+[all:vars]
+ansible_user=wayne
+k3s_version=v1.29.4+k3s1
+cluster_vip=192.168.50.251
