+++ app/globals.css
@@ -111,6 +111,11 @@ h3 { @apply text-2xl; }
   animation: countdown-pill-spin 8s linear infinite;
 }
 
+@keyframes countdown-pill-spin {
+  from { transform: rotate(0deg); }
+  to { transform: rotate(360deg); }
+}
+
 /* Hide horizontal scrollbar track for the sermon rail */
 .hide-scroll::-webkit-scrollbar{ display:none; }
 .hide-scroll{ -ms-overflow-style:none; scrollbar-width:none; }
