+++ tests/basic.test.js
@@ -1,0 +1,6 @@
+import test from 'node:test';
+import assert from 'node:assert/strict';
+
+test('basic truth', () => {
+  assert.equal(1 + 1, 2);
+});
